Improve delete action section of 'table' elements.
Ensure loading of the file won't be interrupted when property
'DELETEACTION' is missing. This is meant to be a tolerance feature, since
it's a mandatory property and therefore, should never occur. While
testing this issue with empiric approach, for some reason it took part
some times in numerous xpo files. At current state, we consider this to
be a kernel bug.